refactor cookies constants to proper directory
